practica02/requirementsFlask.txt
Pillow
Flask
pysvg-py3